pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "shiny>=1.5.1",
    "shinyswatch>=0.7.0",
    "plotly>=5.24.0",
    "shinywidgets>=0.7.0",
    "sqlparse>=0.5.5",
    "accelerate",
    "flask>=3.0.0",
    "hvplot>=0.12.1",
    "jupyter-bokeh>=4.0.5",
    "numpy>=2.3.5",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "toon-py>=0.1.2",
    "torch",
    "transformers==4.53.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["app*"]

[project.scripts]
text2sql = "app.cli.main:main"

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "autoawq==0.2.9",
    "bitsandbytes>=0.48.2",
]

